requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
python-dotenv==1.0.1
Werkzeug==3.0.3
requests==2.32.3
msal==1.29.0
httpx==0.27.0
waitress==3.0.0